cli_monitor: add a 'gdb' command
This patch adds a 'gdb' command to 'cli_monitor', which makes it possible
to debug an application with GDB.

The command works similarly to the 'start' command, but instead of
starting the subsystem binary directly, an 'init' subsystem gets
started, which then starts 'terminal_crosslink', 'noux', 'gdb' and
'gdb_monitor' (which starts the application binary as its target).

So, for the 'gdb' command to work, these additional components need to
be provided as ROM modules, too. The 'gdb' client needs to be
available in a tar archive and the target binary needs to be available
both in raw form (loaded by 'gdb_monitor') as well as in a tar archive
(loaded by 'noux'). Additionally, the source code of the target
application can be provided in a tar archive <target_binary_name>-src.tar.

The patch includes an 'expect' include file (ports/run/noux_gdb.inc)
which provides functions that help to create the needed files:

- 'create_gdb_tar' creates the tar archive for the 'gdb' client
- 'create_binary_tar' creates the tar archive for the target application
- 'create_source_tar' creates the tar archive for the source code of
  the target application
- 'create_binary_and_source_tars' is a convenience wrapper for the previous
  two functions

The patch also includes an example run script
(ports/run/noux_gdb_dynamic.run).

The 'gdb' command supports the following command line options:

- --ram: the initial RAM quota provided to the whole subsystem
         (including the GDB-related components)
- --ram-limit: limit for expanding RAM quota
- --gdb-ram-preserve: the RAM quota that 'gdb_monitor' ahould preserve
    for itself

Fixes 928.