floating window layouter: determine hovered window
The hover report alone is not always reliable enough to determine the
currently hovered window. With this patch, the currently hovered window
is determined by the location of the mouse press event. Unfortunately,
this does not work when child views are involved, since the layouter
does not know about them.