Turmvilla scenario
The scenario, conceived the Hack'n'Hike 2015, builds upon cproc's
recipe for setting up dual-boot system that executes VirtualBox either
on Linux or Genode. It extends the Genode variant with the ability to
interactively start, stop, and edit subsystems.

A shared folder (named "shared") can be used to exchange data between the
Guest OS and the Genode world. On the Noux environment, the shared
folder is available at '/shared/'. The Genode partition is available at
'/genode/'. The LOG messages of higher-level components are written to
'/shared/log'.

After executing the run script, copy the content of the run directory
to the Genode partition and adjust the UUIDs in 'genode/vm_linux.vbox
according to your VDI files.