fixup "lxip: VFS plugin"
vfs lxip: dissolve handles on directory destruct

Each file tracks the handles that exist for it. When the directory of
the file gets destructed it dissolves the handles of all contained files.
This ensures that we do not access an already invalid socket the next time
'poll_all' gets called.