fixup "os/timer: interpolate time via timestamps" (move to base lib)
To enable us to use the timeout lib as back end in the
Timer::Connection, it must be a dependency of the base lib to also be
available in statically linked binaries. As the base lib is a dependency
of the LDSO lib too, we can remove the timeout dependency from the LDSO
lib.

Ref #2400