fixup "os/timer: interpolate time via timestamps" (Timer to Timer::Connection)
There is no use case for having separate classes Timer::Connection and
Genode::Timer in the future. The functionality of Genode::Timer always
needs a Timer::Connection and the Timer::Connection shall only be used
in combination with the timeout framework in the future.

However, we want to stay backwards compatible for now, but the new
non-blocking timeout interface and the old (potentially) blocking
interface of the Timer::Connection can't be used in parallel. Thus, a
Timer::Connection has two modes now. Initially it is in the "legacy"
mode with blocking operations enabled. As soon as the new timeout
interface is used for the first time, it permanently switches to the
"timeout" mode. Attempts to use the legacy interface in the "timeout"
mode cause an exception.

Ref #2400