fixup "os/timer: interpolate time via timestamps" (os/timeout.h to timer/timeout.h)
The timeout.h header is merely for sharing code between
timer_session/connection.h and drivers/timer. Thus, its interface shall
not be used directly in user applications but only through the interface
in timer_session/connection.h. To prevent direct use, we move it to a
public but timer-driver-specific include path.

Ref #2400